Use div macros for header/footer inclusion
In this changeset I removed <div> from allowed tags as it makes it
possible to post arbitrary <div> tags in journals with @<div>@ markup.

It was originally put there to avoid issue with <p> tag wrapping of
included header/footer blocks by macros.

I found a workaround in current Textile parser which allows to disable
wrapping of macros. This can be done by adding "<code></code>" in the
beginning of the line. This way the rest of the line is rendered as is
without being put in a paragraph.

This behavior may change in RedCloth4, but Redmine is still not there so
we can use this workaround for a while, before Redmine comes up with an
appropriate way to disable paragraphs.

fixes https://github.com/haru/redmine_wiki_extensions/issues/8
related https://code.google.com/archive/p/redminewikiext/issues/4
related https://redmine.org/issues/29588
related https://redmine.org/issues/13695